requirements.txt
tensorflow==1.13.1
numpy
opencv-python
flask
pillow
